Rename Build namespace to Rebuild to fix collision
The Aspire.Hosting.Build namespace collided with the Build type in
Aspire.Hosting.Docker/Resources/ComposeNodes/Service.cs, causing
CS0118: 'Build' is a namespace but is used like a type.

Renamed to Aspire.Hosting.Rebuild to avoid the conflict.
