+++ src/Aspire.Hosting/DistributedApplicationBuilder.cs
@@ -634,7 +634,7 @@ public class DistributedApplicationBuilder : IDistributedApplicationBuilder
         {
             _innerBuilder.Services.AddSingleton<ResourceHealthCheckService>();
             _innerBuilder.Services.AddHostedService<ResourceHealthCheckService>(sp => sp.GetRequiredService<ResourceHealthCheckService>());
-            _innerBuilder.Services.AddHostedService<Build.ProjectChangeDetectionService>();
+            _innerBuilder.Services.AddHostedService<Rebuild.ProjectChangeDetectionService>();
         }
     }
