+++ tests/Aspire.Hosting.Tests/Rebuild/ProjectFileClosureTests.cs
@@ -1,9 +1,9 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Aspire.Hosting.Build;
+using Aspire.Hosting.Rebuild;
 
-namespace Aspire.Hosting.Tests.Build;
+namespace Aspire.Hosting.Tests.Rebuild;
 
 public class ProjectFileClosureTests
 {
